refactor(server): split match.ts — extract handleGetMatchState (S27.8.18)
Third slice on `apps/server/src/routes/match.ts`. Extract
`handleGetMatchState` (~221 lines) into `routes/match-state-handler.ts`
(285 lines).

The handler covers `GET /:id/state` and is the largest remaining
handler in `match.ts`:
- Atomic idle → setup transition (`ensureSetupPhasePersisted`)
- AI setup trigger (`runAISetupIfNeeded`)
- Canonical `gameState` selection by `match.status`
  (prematch/prematch-setup/active)
- Prematch reconstruction via `setupPreMatchWithTeams` (dynamic import
  from `@bb/game-engine`)
- Response enrichment for active/prematch-setup (currentTeam,
  isMyTurn, myTeamSide, moveCount, lastMoveAt)

Re-export from `match.ts` to preserve the public API consumed by
integration tests.

Leaf dependencies only. No cycle back to `match.ts`.

`match.ts` shrinks 1393 -> 1176 lines (-217 net lines, cumulative
-590 vs origin 1766 since S27.8.16).

- 1787/1787 server tests pass (no regression)
- typecheck clean
