+++ apps/server/src/routes/match.ts
@@ -670,230 +670,13 @@ export async function handleGetMatchSummary(
 
 router.get("/:id/summary", authUser, handleGetMatchSummary);
 
-// Nouvel endpoint pour l'état du jeu (S25.5l — ApiResponse<T>)
-export async function handleGetMatchState(
-  req: AuthenticatedRequest,
-  res: Response,
-): Promise<void> {
-  try {
-    const matchId = req.params.id;
+// S27.8.18 — `handleGetMatchState` extrait dans
+// `routes/match-state-handler.ts`. Re-export pour preserver l'API
+// publique consommee par les tests d'integration.
+export { handleGetMatchState } from './match-state-handler';
+import { handleGetMatchState as handleGetMatchStateImpl } from './match-state-handler';
 
-    // Persist the idle → setup transition atomically before reading the
-    // match. This makes the rest of this handler a pure read: any concurrent
-    // call sees the same persisted state instead of running the transition
-    // again in-memory. The returned `gameState` is the canonical post-bypass
-    // state (preferring `setup-init`/`validate-setup` over a stale
-    // `pre-match-sequence` written by the parallel automation).
-    const ensured = await ensureSetupPhasePersisted(matchId, prisma as any);
-
-    // Once the setup phase is persisted, if the AI is the current coach,
-    // place its 11 players automatically so the match does not hang on a
-    // coach that has no client submitting on its behalf.
-    if (ensured?.gameState?.preMatch?.phase === "setup") {
-      await runAISetupIfNeeded(matchId, prisma as any);
-    }
-
-    const match = await prisma.match.findUnique({
-      where: { id: matchId },
-      include: { turns: { orderBy: { number: "asc" } } },
-    });
-    if (!match) {
-      sendError(res, "Partie introuvable", 404);
-      return;
-    }
-
-    if (match.status === "pending") {
-      sendError(res, "Partie pas encore prête", 400);
-      return;
-    }
-
-    let gameState: any;
-
-    // Pour prematch-setup, chercher le dernier turn avec gameState
-    // (validate-setup, pre-match-sequence, inducements, ou coin-toss)
-    const startTurn = match.turns.find(
-      (t: any) =>
-        t.payload?.type === "start" || t.payload?.type === "coin-toss",
-    );
-    if (match.status === "prematch-setup" || match.status === "active") {
-      // L'état canonique est TOUJOURS le dernier turn avec gameState : il
-      // intègre les étapes post-setup (place-kickoff-ball, kick-deviation,
-      // resolve-kickoff-event, gameplay-move). Préférer un `validate-setup`
-      // plus ancien renvoyait un état figé au client après que la séquence
-      // kickoff ait progressé côté serveur, bloquant le joueur sur l'écran
-      // "En attente du placement du ballon".
-      const latestStateTurn = [...match.turns]
-        .reverse()
-        .find((t: any) => t.payload?.gameState);
-      if (latestStateTurn) {
-        const gs = latestStateTurn.payload.gameState;
-        gameState = typeof gs === "string" ? JSON.parse(gs) : gs;
-      } else if (ensured) {
-        gameState = ensured.gameState;
-      } else if (startTurn) {
-        const gs = startTurn.payload.gameState;
-        gameState = typeof gs === "string" ? JSON.parse(gs) : gs;
-      }
-    } else if (startTurn) {
-      const gs = startTurn.payload.gameState;
-      gameState = typeof gs === "string" ? JSON.parse(gs) : gs;
-    } else if (match.status === "prematch") {
-      // Reconstruire si pas encore de turn start (cas edge)
-      const selections = await prisma.teamSelection.findMany({
-        where: { matchId },
-        include: {
-          user: true,
-          teamRef: true,
-        },
-        orderBy: { createdAt: "asc" },
-      });
-      if (selections.length < 2) {
-        sendError(res, "Équipes pas prêtes", 400);
-        return;
-      }
-
-      const [s1, s2] = selections;
-
-      // Fetch teams séparément pour players
-      const teamA = s1.teamId
-        ? await prisma.team.findUnique({
-            where: { id: s1.teamId },
-            include: { players: true },
-          })
-        : null;
-      const teamB = s2.teamId
-        ? await prisma.team.findUnique({
-            where: { id: s2.teamId },
-            include: { players: true },
-          })
-        : null;
-
-      if (!teamA || !teamB) {
-        serverLog.log("Teams not found:", s1.teamId, s2.teamId);
-        sendError(res, "Équipes non trouvées", 400);
-        return;
-      }
-
-      const teamAData = teamA.players
-        .filter((p: any) => !p.dead)
-        .map((p: any) => ({
-          id: p.id,
-          name: p.name,
-          position: p.position,
-          number: p.number,
-          ma: p.ma,
-          st: p.st,
-          ag: p.ag,
-          pa: p.pa,
-          av: p.av,
-          skills: p.skills || "",
-        }));
-      const teamBData = teamB.players
-        .filter((p: any) => !p.dead)
-        .map((p: any) => ({
-          id: p.id,
-          name: p.name,
-          position: p.position,
-          number: p.number,
-          ma: p.ma,
-          st: p.st,
-          ag: p.ag,
-          pa: p.pa,
-          av: p.av,
-          skills: p.skills || "",
-        }));
-
-      serverLog.log(
-        "Players loaded for prematch:",
-        teamAData.length,
-        teamBData.length,
-      ); // Trace
-
-      const teamAName = s1.teamRef?.name || s1.team || "Team A";
-      const teamBName = s2.teamRef?.name || s2.team || "Team B";
-
-      const { setupPreMatchWithTeams } = await import("@bb/game-engine");
-      // H.6 — propagate roster slugs so the client renderer can pick per-roster colors.
-      gameState = setupPreMatchWithTeams(
-        teamAData,
-        teamBData,
-        teamAName,
-        teamBName,
-        {
-          teamARoster: teamA.roster,
-          teamBRoster: teamB.roster,
-        },
-      );
-    } else {
-      // Pour active, dernier turn
-      const lastTurn = match.turns[match.turns.length - 1];
-      if (!lastTurn?.payload?.gameState) {
-        sendError(res, "État non trouvé", 500);
-        return;
-      }
-      const gs = lastTurn.payload.gameState;
-      gameState = typeof gs === "string" ? JSON.parse(gs) : gs;
-    }
-
-    // Pour les matchs actifs, enrichir la réponse avec des métadonnées
-    if (match.status === "active" && gameState) {
-      const userTeamSide = await getUserTeamSide(
-        prisma as any,
-        matchId,
-        req.user!.id,
-      );
-      const phase = gameState.preMatch?.phase;
-      const isMyTurn = userTeamSide
-        ? phase === "setup" || phase === "kickoff-sequence"
-          ? gameState.preMatch?.currentCoach === userTeamSide
-          : gameState.currentPlayer === userTeamSide
-        : false;
-      const moveCount = match.turns.filter(
-        (t: any) => t.payload?.type === "gameplay-move",
-      ).length;
-      const lastMoveTurn = [...match.turns]
-        .reverse()
-        .find((t: any) => t.payload?.type === "gameplay-move");
-
-      sendSuccess(res, {
-        gameState,
-        matchStatus: match.status,
-        currentTeam: gameState.currentPlayer,
-        isMyTurn,
-        myTeamSide: userTeamSide,
-        moveCount,
-        lastMoveAt: lastMoveTurn?.createdAt || null,
-      });
-      return;
-    }
-
-    // Pure read for prematch-setup: the idle → setup transition was already
-    // persisted by ensureSetupPhasePersisted at the top of this handler.
-    if (match.status === "prematch-setup") {
-      const userTeamSide = await getUserTeamSide(
-        prisma as any,
-        matchId,
-        req.user!.id,
-      );
-      sendSuccess(res, {
-        gameState,
-        matchStatus: match.status,
-        myTeamSide: userTeamSide,
-        isMyTurn: userTeamSide
-          ? gameState?.preMatch?.currentCoach === userTeamSide
-          : false,
-      });
-      return;
-    }
-
-    sendSuccess(res, { gameState });
-  } catch (e: unknown) {
-    serverLog.error(e);
-    sendError(res, "Erreur serveur", 500);
-  }
-}
-
-router.get("/:id/state", authUser, handleGetMatchState);
+router.get("/:id/state", authUser, handleGetMatchStateImpl);
 
 // Valider le placement des joueurs en phase setup
 router.post(
